:sparkles: feat: implement SummaryPage component for displaying study session details
